move Rate class to rates data file
